rename: component 폴더 위치 이동
